backend/MIHIRESS/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.112.1
Flask==3.0.3
flatbuffers==24.3.25
gast==0.6.0
google-ai-generativelanguage==0.6.9
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.1
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
h5py==3.11.0
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.5.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.0
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
patsy==0.5.6
proto-plus==1.24.0
protobuf==5.28.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rich==13.7.1
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.0
setuptools==72.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.2
statsmodels==0.14.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.4.0
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
Werkzeug==3.0.3
wheel==0.44.0
wrapt==1.16.0